pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datahub-talk-to-data"
version = "0.1.0"
description = "LangGraph-based talk-to-data agent with DataHub context and pluggable query engines"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # DB
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "aiomysql>=0.2.0",
    # Config
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0",
    # Agent
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    # DataHub context
    "datahub-agent-context[langchain]>=1.5.0",
    "acryl-datahub[base]>=1.5.0",
    # Query engines
    "snowflake-connector-python[pandas,secure-local-storage]>=3.10.0",
    "pymysql>=1.1.0",
    # Observability (OTEL) — no-op unless OTEL_EXPORTER_OTLP_ENDPOINT is set
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-langchain>=0.32.1",
    # Utilities
    "orjson>=3.10.0",
    "jsonschema>=4.23.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "keyring>=25.7.0",
    "langchain-google-genai>=4.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "types-pyyaml",
    "types-pymysql",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/src/analytics_agent"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module-level import not at top (intentional in main.py for load_dotenv)
    "B008",   # function call in default args (FastAPI Depends() pattern)
    "B027",   # empty method in abstract class (aclose in QueryEngine)
    "B904",   # raise from err in except (style preference)
    "B905",   # zip without strict (not needed for our use cases)
    "SIM105", # contextlib.suppress (try/except/pass is clearer here)
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]
